move meadows into valley (and added npm commands)
