+++ .gitignore
@@ -1,3 +1,3 @@
-meadows/
+valley/
 node_modules/
 .DS_Store
