+++ README.md
@@ -1,5 +1,17 @@
 ## Basic usage:
-Modify meadows.js with the paths to the dotfiles you want to back up. Then run `node ./gather.js` to back the files up in `meadows/**` locally. Run `git init` from `meadows/`, then push your config files to git. From there, clone wildflower and your meadows repo. Run `node ./sow.js` to place the files in the correct place.
+### First time users
+
+1. Clone this repo.
+2. Run `npm run till`
+3. Modify the created `./valley/meadows.js` with the paths of config files to save.
+4. Run `npm run gather` to collect all the files you want to save.
+5. Store the contents of `./valley` any way you like (recommendation is git).
+
+### Returning users 
+
+1. Clone this repo.
+2. Copy your existing `./valley` into this repo.
+3. Run `npm run sow` to take the files in your `valley` folder and distribute them
 
 ## Known bugs:
 - Gather deletes a directory synchronously when an async workflow would be ideal
