+++ package.json
@@ -4,7 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "till": "node ./till.js",
+    "gather": "node ./gather.js",
+    "sow": "node ./sow.js"
   },
   "author": "",
   "license": "ISC",
